Began work on creating base notebooks for testing
